requirements.txt
Django==4.2.5
mysqlclient==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==21.2.0
